requirements.txt
requests
textwrap
python-dotenv
pylint
autopep8
pytest
pytest-cov